refactor: 인수테스트 클래스 통합
